+++ src/components/footer.tsx
@@ -1,103 +1,11 @@
 import React from 'react'
 import { ListItem } from './list-item'
-import { Info } from '@phosphor-icons/react'
+
 import clsx from 'clsx'
 import { GetSupport } from './get-support'
+import { footerLinks } from '../config'
 
 export function Footer({ className }: { className?: string }) {
-  const footerLinks = [
-    {
-      title: 'Build with ink!',
-      links: [
-        {
-          label: 'Docs',
-          href: '/docs',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Tutorials',
-          href: 'https://github.com/use-ink/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Tooling',
-          href: 'https://github.com/use-ink/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Chains',
-          href: 'https://github.com/use-ink/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'YouTube',
-          href: 'https://www.youtube.com/@ink_lang',
-          icon: <Info size={20} />,
-        },
-      ],
-    },
-    {
-      title: 'Join our community',
-      links: [
-        {
-          label: 'About',
-          href: 'https://discord.gg/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Bounty',
-          href: 'https://twitter.com/ink_lang',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Hackathons',
-          href: 'https://discord.gg/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Stories',
-          href: 'https://discord.gg/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Support',
-          href: 'https://discord.gg/ink',
-          icon: <Info size={20} />,
-        },
-      ],
-    },
-    {
-      title: 'Helpful Links',
-      links: [
-        {
-          label: 'ink!ubator',
-          href: 'https://discord.gg/ink',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'ink! examples',
-          href: 'https://github.com/use-ink/ink-examples',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Funding Programs',
-          href: 'https://github.com/use-ink/ink-docs',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'Brand Assets',
-          href: 'https://github.com/use-ink/ink-docs',
-          icon: <Info size={20} />,
-        },
-        {
-          label: 'FAQ',
-          href: 'https://github.com/use-ink/ink-docs',
-          icon: <Info size={20} />,
-        },
-      ],
-    },
-  ]
-
   // linear-gradient(180deg,rgba(33,15,59,0) 0%,rgb(29,24,61) 38%,rgb(0,0,0) 100%)
 
   return (