+++ src/pages/index.tsx
@@ -2,16 +2,13 @@ import React from 'react'
 import Link from '@docusaurus/Link'
 
 import '../css/landing.css'
-import LandingNav from '../components/landing/nav'
-
+import { Navbar } from '../components/landing/nav'
+import { Footer } from '../components/footer'
 export default function Home() {
   return (
     <>
-      <LandingNav />
-      <div
-        className="flex flex-col items-center justify-center h-screen dark:bg-gradient-to-b dark:from-purple-900 dark:via-black dark:to-purple-400"
-        style={{ fontFamily: 'Montserrat  ' }}
-      >
+      <Navbar />
+      <div className="flex flex-col items-center justify-center h-screen" style={{ fontFamily: 'Montserrat  ' }}>
         <h1 className="text-4xl font-bold">Hello ink!</h1>
         <p className="text-sm text-gray-500">(This will be the revamped ink! landing page)</p>
         <div className="flex flex-row gap-4">
@@ -23,13 +20,8 @@ export default function Home() {
           </Link>
         </div>
       </div>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
-      <section className="h-screen bg-purple-950"> </section>
+      <section className="h-screen"> </section>
+      <Footer />
     </>
   )
 }