+++ src/theme/DocsRoot/index.tsx
@@ -4,6 +4,8 @@ import { ThemeClassNames, HtmlClassNameProvider } from '@docusaurus/theme-common
 import renderRoutes from '@docusaurus/renderRoutes'
 import Layout from '@theme/Layout'
 
+import '../../css/docs.css'
+
 import type { Props } from '@theme/DocVersionRoot'
 
 export default function DocsRoot(props: Props): ReactNode {