+++ CHANGELOG.md
@@ -31,6 +31,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - repo structure
 
+  - create `shared.cs`, `docs.css`, `landing.css`
+
 - misc
   - update github workflow to use node 18 -> 20
   - make docusuaurus.config.js => docusuaurus.config.ts and fix duplicate config entries